requirements.txt
discord.py
requests
beautifulsoup4
feedparser